# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SalonFlow API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Service {
  id: number;
  name: string;
  durationMinutes: number;
  price: number;
  description?: string | null;
  createdAt: string;
}

export interface CreateServiceBody {
  name: string;
  durationMinutes: number;
  price: number;
  description?: string | null;
}

export interface Staff {
  id: number;
  name: string;
  role: string;
  createdAt: string;
}

export interface CreateStaffBody {
  name: string;
  role: string;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Booking {
  id: number;
  referenceNumber: string;
  serviceId: number;
  staffId?: number | null;
  customerName: string;
  customerPhone: string;
  customerNotes?: string | null;
  bookingDate: string;
  bookingTime: string;
  status: BookingStatus;
  adminNotes?: string | null;
  totalPrice: number;
  createdAt: string;
  service?: Service | null;
  staff?: Staff | null;
}

export interface CreateBookingBody {
  serviceId: number;
  staffId?: number | null;
  customerName: string;
  customerPhone: string;
  customerNotes?: string | null;
  bookingDate: string;
  bookingTime: string;
}

export type UpdateBookingBodyStatus =
  (typeof UpdateBookingBodyStatus)[keyof typeof UpdateBookingBodyStatus];

export const UpdateBookingBodyStatus = {
  pending: "pending",
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateBookingBody {
  status?: UpdateBookingBodyStatus;
  adminNotes?: string | null;
}

export interface AvailableSlotsResponse {
  date: string;
  slots: string[];
  isClosed: boolean;
}

export interface AdminLoginBody {
  username: string;
  password: string;
}

export interface AdminLoginResponse {
  success: boolean;
  token: string;
}

export interface AdminStats {
  totalBookings: number;
  pendingBookings: number;
  confirmedBookings: number;
  completedBookings: number;
  cancelledBookings: number;
  todayBookings: number;
  totalRevenue: number;
}

export type ListBookingsParams = {
  status?: ListBookingsStatus;
};

export type ListBookingsStatus =
  (typeof ListBookingsStatus)[keyof typeof ListBookingsStatus];

export const ListBookingsStatus = {
  pending: "pending",
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type GetAvailableSlotsParams = {
  /**
   * ISO date string (YYYY-MM-DD)
   */
  date: string;
  serviceId: number;
  staffId?: number;
};