# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SalonFlow API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: services
    description: Salon services
  - name: staff
    description: Salon staff
  - name: bookings
    description: Booking operations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all services
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create a new service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceBody"
      responses:
        "201":
          description: Created service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"

  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get a service by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found
    put:
      operationId: updateService
      tags: [services]
      summary: Update a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceBody"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List all staff
      responses:
        "200":
          description: List of staff members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Staff"
    post:
      operationId: createStaff
      tags: [staff]
      summary: Create a staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffBody"
      responses:
        "201":
          description: Created staff member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"

  /staff/{id}:
    get:
      operationId: getStaff
      tags: [staff]
      summary: Get a staff member by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Staff member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"
        "404":
          description: Not found
    put:
      operationId: updateStaff
      tags: [staff]
      summary: Update a staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffBody"
      responses:
        "200":
          description: Updated staff member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"
    delete:
      operationId: deleteStaff
      tags: [staff]
      summary: Delete a staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List all bookings (admin)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, confirmed, completed, cancelled]
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking (customer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Created booking with reference number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "409":
          description: Time slot conflict

  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a booking by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Update booking status and admin notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingBody"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  /bookings/available-slots:
    get:
      operationId: getAvailableSlots
      tags: [bookings]
      summary: Get available time slots for a given date, service and optional staff
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            description: ISO date string (YYYY-MM-DD)
        - name: serviceId
          in: query
          required: true
          schema:
            type: integer
        - name: staffId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Available time slots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailableSlotsResponse"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Invalid credentials

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get booking stats for admin dashboard
      responses:
        "200":
          description: Admin stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Service:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        durationMinutes:
          type: integer
        price:
          type: number
        description:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - name
        - durationMinutes
        - price
        - createdAt

    CreateServiceBody:
      type: object
      properties:
        name:
          type: string
        durationMinutes:
          type: integer
        price:
          type: number
        description:
          type: string
          nullable: true
      required:
        - name
        - durationMinutes
        - price

    Staff:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - role
        - createdAt

    CreateStaffBody:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
      required:
        - name
        - role

    Booking:
      type: object
      properties:
        id:
          type: integer
        referenceNumber:
          type: string
        serviceId:
          type: integer
        staffId:
          type: integer
          nullable: true
        customerName:
          type: string
        customerPhone:
          type: string
        customerNotes:
          type: string
          nullable: true
        bookingDate:
          type: string
        bookingTime:
          type: string
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        adminNotes:
          type: string
          nullable: true
        totalPrice:
          type: number
        createdAt:
          type: string
        service:
          $ref: "#/components/schemas/Service"
          nullable: true
        staff:
          $ref: "#/components/schemas/Staff"
          nullable: true
      required:
        - id
        - referenceNumber
        - serviceId
        - customerName
        - customerPhone
        - bookingDate
        - bookingTime
        - status
        - totalPrice
        - createdAt

    CreateBookingBody:
      type: object
      properties:
        serviceId:
          type: integer
        staffId:
          type: integer
          nullable: true
        customerName:
          type: string
        customerPhone:
          type: string
        customerNotes:
          type: string
          nullable: true
        bookingDate:
          type: string
        bookingTime:
          type: string
      required:
        - serviceId
        - customerName
        - customerPhone
        - bookingDate
        - bookingTime

    UpdateBookingBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        adminNotes:
          type: string
          nullable: true

    AvailableSlotsResponse:
      type: object
      properties:
        date:
          type: string
        slots:
          type: array
          items:
            type: string
        isClosed:
          type: boolean
      required:
        - date
        - slots
        - isClosed

    AdminLoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    AdminLoginResponse:
      type: object
      properties:
        success:
          type: boolean
        token:
          type: string
      required:
        - success
        - token

    AdminStats:
      type: object
      properties:
        totalBookings:
          type: integer
        pendingBookings:
          type: integer
        confirmedBookings:
          type: integer
        completedBookings:
          type: integer
        cancelledBookings:
          type: integer
        todayBookings:
          type: integer
        totalRevenue:
          type: number
      required:
        - totalBookings
        - pendingBookings
        - confirmedBookings
        - completedBookings
        - cancelledBookings
        - todayBookings
        - totalRevenue