# lib/api-zod/src/generated/types/availableSlotsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SalonFlow API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AvailableSlotsResponse {
  date: string;
  slots: string[];
  isClosed: boolean;
}