# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SalonFlow API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from "./bookingStatus";
import type { Service } from "./service";
import type { Staff } from "./staff";

export interface Booking {
  id: number;
  referenceNumber: string;
  serviceId: number;
  staffId?: number | null;
  customerName: string;
  customerPhone: string;
  customerNotes?: string | null;
  bookingDate: string;
  bookingTime: string;
  status: BookingStatus;
  adminNotes?: string | null;
  totalPrice: number;
  createdAt: string;
  service?: Service | null;
  staff?: Staff | null;
}