# lib/api-zod/src/generated/types/createBookingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SalonFlow API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBookingBody {
  serviceId: number;
  staffId?: number | null;
  customerName: string;
  customerPhone: string;
  customerNotes?: string | null;
  bookingDate: string;
  bookingTime: string;
}